header and footer changes
